% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgPHYLset-methods.R
\name{ctgPHYLset-methods}
\alias{ctgPHYLset-methods}
\alias{cellscapeData}
\alias{ctgPHYLset,ANY,ANY-method}
\alias{cellscapeData<-}
\alias{timescapeData}
\alias{timescapeData<-}
\alias{mapscapeData}
\alias{mapscapeData<-}
\alias{treeList}
\alias{treeList<-}
\alias{originalTrees}
\alias{originalTrees<-}
\title{Methods for the ctgPHYLset class}
\usage{
cellscapeData(cs)

cellscapeData(cs, pt) <- value

timescapeData(cs)

timescapeData(cs, pt) <- value

mapscapeData(cs)

mapscapeData(cs, pt) <- value

treeList(cs)

treeList(cs, tt) <- value

originalTrees(cs)

originalTrees(cs, tt) <- value
}
\arguments{
\item{cs}{A ctgPHYLset object}

\item{pt}{The name of the \pkg{cellscape}, \pkg{timescape}, or
\pkg{mapscape} parameter to store}

\item{value}{\code{cellscapeData(cs, pt)<-}: The value to use as a named parameter for
cellscape, used by \code{generate_tree(treeType = "cellscape")} in the
\pkg{cellTreeGenerator} workflow

\code{timescapeData(cs, pt)<-}: The value to use as a named parameter for
timescape used by \code{generate_tree(treeType = "timescape")} in the
\pkg{cellTreeGenerator} workflow

\code{mapscapeData(cs)<-}: The value to use as a named parameter for
mapscape, used by \code{generate_tree(treeType = "mapscape")} in the
\pkg{cellTreeGenerator} workflow}

\item{tt}{The type of tree being stored}
}
\value{
An updated ctgPHYLset object, or the contents of a slot of the
ctgPHYLset object
}
\description{
These methods operate on ctgPHYLset objects. Please note that
treeList<- and originalTrees<- are not intended to be
called directly.
}
\examples{
\dontrun{
cellTreeInfo(samplectgPHYLset) <- "Hours"
cti <- cellTreeInfo(samplectgPHYLset)
monocleInfo(samplectgPHYLset) <- c("gene_short_name", "MYF5",
                                     "ANPEP", "FPKM")
mi <- monocleInfo(samplectgPHYLset)
TSCANinfo(samplectgPHYLset) <- "ENSG00000000003.10"
ti <- TSCANinfo(samplectgPHYLset)
oncoNEMdata(samplectgPHYLset) <- my_snp_matrix
od <- oncoNEMdata(samplectgPHYLset)
CanopyData(samplectgPHYLset) <- my_Canopy_Data_list
cd <- CanopyData(samplectgPHYLset)
trees <- treeList(samplectgPHYLset)
originalTrees <- originalTrees(samplectgPHYLset)
}
}
\keyword{internal}
