% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgPHYLset-class.R
\docType{class}
\name{ctgPHYLset}
\alias{ctgPHYLset}
\alias{ctgPHYLset-class}
\title{The ctgPHYLset class}
\description{
The main class used by the \pkg{ctgPHYL} packages to hold PyClone output
data and generated tree results.
}
\details{
Methods that operate on ctgPHYLset objectsvcomprise the ctgPHYL workflow.
}
\section{Fields}{

\describe{
\item{\code{cellscapeData}}{A list containing the data to be used by
\code{generate_tree(treeType = "cellscape")} in the
\pkg{ctgPHYL} workflow}

\item{\code{timescapeData}}{A list containing the data to be used by
\code{generate_tree(treeType = "timescape")} in the
\pkg{ctgPHYL} workflow}

\item{\code{mapscapeData}}{A list containing the data to be used by
\code{generate_tree(treeType = "mapscape")} in the
\pkg{ctgPHYL} workflow}

\item{\code{treeList}}{A list containing the simplified igraph representation of
the trees generated by the \pkg{ctgPHYL} workflow}

\item{\code{originalTrees}}{A list containing the trees generated by the
\pkg{ctgPHYL} workflow, in their original formats}
}}

