% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCellscape.R
\name{makeCellscape}
\alias{makeCellscape}
\title{A Cell Tree Generating Function using cellscape}
\usage{
makeCellscape(dataSet)
}
\arguments{
\item{dataSet}{a ctgPHYLset object}
}
\value{
an updated ctgPHYLset object
}
\description{
This function, called by \code{\link{generate_tree}}, creates visualizations
using the data and \pkg{cellscape}.
}
\keyword{internal}
