% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newCtgPHYLset.R
\name{newctgPHYLset}
\alias{newctgPHYLset}
\title{Creates a new ctgPHYLset object.}
\usage{
newctgPHYLset()
}
\value{
a new, empty ctgPHYLset object to store parameters, inputs, and
outputs
}
\description{
Creates a new ctgPHYLset object.
}
\examples{
# create example ctgPHYLset
ctgPHYLset <- newctgPHYLset()
}
