% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOriginalTree.R
\name{plotOriginalTree}
\alias{plotOriginalTree}
\title{Display Original ctgPHYL Plots}
\usage{
plotOriginalTree(dataSet, treeType)
}
\arguments{
\item{dataSet}{a ctgPHYLset object}

\item{treeType}{the type of tree to display.  Must be one of
\code{names(originalTrees(dataSet))}}
}
\value{
An object of class \code{c(<tree type>, "htmlwidget")}.
}
\description{
Displays the original plots created by the ctgPHYL package and
stored in the \link{originalTrees} slot of a ctgPHYLset object.
}
\examples{
# load data for cellscape
cs_tree_edges <- read.csv(system.file("extdata", "cnv_tree_edges.csv",
                         package = "cellscape"))
cs_cnv_data <- read.csv(system.file("extdata", "cnv_data.csv",
                                     package = "cellscape"))
cs_sc_annot <- read.csv(system.file("extdata", "cnv_annots.tsv",
                                    package = "cellscape"), sep="\t")
cs_clone_colours <- data.frame(clone_id = c("1","2","3"),
                               colour = c("7fc97f", "beaed4", "fdc086"))
# create example ctgPHYLset and load data into it
ctgPHYLset <- ctgPHYLset()
cellscapeData(ctgPHYLset, "clone_colours") <- cs_clone_colours
cellscapeData(ctgPHYLset, "tree_edges") <- cs_tree_edges
cellscapeData(ctgPHYLset, "cnv_data") <- cs_cnv_data
cellscapeData(ctgPHYLset, "sc_annot") <- cs_sc_annot

# run generate_tree()
ctgPHYLset <- generate_tree(dataSet = ctgPHYLset, treeType = "cellscape")

# view names of original trees
names(originalTrees(ctgPHYLset))

# plot original trees
plotOriginalTree(ctgPHYLset, "cellscape")
}
